# docs/api-spec.yml
openapi: "3.1.0"
info:
  title: Agente IA Autónomos — MVP P04 API
  version: "1.0.0"
  description: |
    API backend for the autonomous tax agent MVP.
    Process P04: IVA Trimestral (Modelo 303).

    All endpoints require a valid Supabase JWT in the Authorization header.
    Row Level Security is enforced at the database level — users can only
    access their own data.

servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://api.autonomos-ia.com
    description: Production

security:
  - supabaseAuth: []

components:
  securitySchemes:
    supabaseAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase Auth JWT token

  schemas:
    Error:
      type: object
      required: [error, detail, proceso]
      properties:
        error:
          type: string
          example: "PRESENTACION_DUPLICADA"
        detail:
          type: string
          example: "Ya existe una presentación para 1T 2026"
        proceso:
          type: string
          example: "P04"

    FacturaEmitidaInput:
      type: object
      required: [numero_factura, fecha, base_imponible, tipo_iva]
      properties:
        numero_factura:
          type: string
          example: "F2026-001"
        fecha:
          type: string
          format: date
          example: "2026-03-15"
        nif_cliente:
          type: string
          nullable: true
          example: "B12345678"
        nombre_cliente:
          type: string
          nullable: true
          example: "Empresa Cliente SL"
        base_imponible:
          type: number
          format: decimal
          minimum: 0
          example: 1000.00
        tipo_iva:
          type: integer
          enum: [0, 4, 10, 21]
          example: 21
        retencion_irpf:
          type: number
          format: decimal
          default: 0
          example: 15.0
        es_isp:
          type: boolean
          default: false
        cobrada:
          type: boolean
          default: true

    FacturaRecibidaInput:
      type: object
      required: [fecha, base_imponible, categoria_gasto, porcentaje_deducible]
      properties:
        fecha:
          type: string
          format: date
        nif_proveedor:
          type: string
          nullable: true
        nombre_proveedor:
          type: string
          nullable: true
        descripcion:
          type: string
          nullable: true
        categoria_gasto:
          type: string
          example: "software_saas"
          description: Must be a valid key from tabla_deducibilidad
        base_imponible:
          type: number
          format: decimal
          minimum: 0
        tipo_iva:
          type: integer
          enum: [0, 4, 10, 21]
          nullable: true
        cuota_iva:
          type: number
          format: decimal
          nullable: true
        porcentaje_deducible:
          type: number
          format: decimal
          minimum: 0
          maximum: 100
        es_bien_inversion:
          type: boolean
          default: false
        es_isp:
          type: boolean
          default: false

    OcrResult:
      type: object
      properties:
        extracted:
          type: object
          description: Fields extracted from the invoice PDF
          properties:
            numero_factura:
              type: object
              properties:
                value: { type: string }
                confidence: { type: number, minimum: 0, maximum: 1 }
            fecha:
              type: object
              properties:
                value: { type: string, format: date }
                confidence: { type: number }
            nif_emisor:
              type: object
              properties:
                value: { type: string }
                confidence: { type: number }
            base_imponible:
              type: object
              properties:
                value: { type: number }
                confidence: { type: number }
            tipo_iva:
              type: object
              properties:
                value: { type: integer }
                confidence: { type: number }
            cuota_iva:
              type: object
              properties:
                value: { type: number }
                confidence: { type: number }
        requires_review:
          type: boolean
          description: True if any field has confidence < 0.8
        pdf_path:
          type: string
          description: Supabase Storage path of uploaded PDF

    ResultadoM303:
      type: object
      properties:
        ejercicio:
          type: integer
          example: 2026
        periodo:
          type: string
          example: "1T"
        total_devengado:
          type: number
          description: Total accrued VAT (IVA repercutido)
          example: 4200.00
        total_deducible:
          type: number
          description: Total deductible VAT (IVA soportado deducible)
          example: 1800.00
        saldo_compensar_anterior:
          type: number
          description: Negative IVA balance carried from previous quarter (casilla 110)
          example: 0.00
        resultado:
          type: number
          description: Final result. Positive = a ingresar. Negative = a compensar.
          example: 2400.00
        tipo_resultado:
          type: string
          enum: [a_ingresar, a_compensar, a_devolver, sin_actividad]
          example: "a_ingresar"
        fecha_limite:
          type: string
          format: date
          description: Filing deadline (adjusted for weekends/holidays)
          example: "2026-04-20"
        casillas:
          type: object
          description: Full M303 casilla map for AEAT form filling
          additionalProperties:
            type: number

    ProcesoEstado:
      type: object
      description: |
        The literal shape of a `presentacion` row plus a computed `qr_url` --
        NOT keyed by `proceso_id` (the thread_id). `id` here is
        `presentacion`'s own server-generated UUID, a different identifier
        than the `proceso_id` path parameter used to look this row up (see
        the `proceso_id` parameter description on GET /estado/{proceso_id}).
      properties:
        id:
          type: string
          format: uuid
        estado:
          type: string
          enum: [pendiente, calculado, confirmado, presentando, presentado, error, cancelado]
        ejercicio:
          type: integer
          example: 2026
        periodo:
          type: string
          example: "1T"
        resultado:
          type: number
          nullable: true
          description: Net M303 result (positive = a ingresar, negative = a compensar/devolver) -- not the full ResultadoM303 breakdown, which is only ever returned by POST /calcular
        csv_aeat:
          type: string
          nullable: true
          description: 16-char AEAT confirmation code
        nrc:
          type: string
          nullable: true
          description: 22-char payment reference (only if a_ingresar)
        error_code:
          type: string
          nullable: true
        error_detail:
          type: string
          nullable: true
        qr_url:
          type: string
          nullable: true
          description: Signed Cl@ve Móvil QR image URL, computed only while estado=presentando; null otherwise or if the QR hasn't been generated yet this attempt

paths:
  /api/proceso/p04/iniciar:
    post:
      summary: Start P04 process
      description: |
        Initializes a new P04 process for the authenticated user.
        Detects the current period (1T/2T/3T/4T) from the current date.
        Fails if a declaration already exists for the detected period
        (returns PRESENTACION_DUPLICADA error).
      operationId: iniciarP04
      responses:
        "200":
          description: Process initialized
          content:
            application/json:
              schema:
                type: object
                properties:
                  proceso_id:
                    type: string
                    description: The LangGraph thread_id ("{user_id}:P04:{ejercicio}:{periodo}"), not a UUID -- pass this back verbatim to every other P04 endpoint
                    example: "3fa8...:P04:2026:1T"
                  ejercicio:
                    type: integer
                    example: 2026
                  periodo:
                    type: string
                    example: "1T"
                  fecha_limite:
                    type: string
                    format: date
                    example: "2026-04-20"
        "409":
          description: Declaration already exists for this period
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /api/proceso/p04/facturas:
    post:
      summary: Manually create an invoice (Phase 1 JSON CRUD, no OCR)
      description: |
        Persists a manually-entered invoice (no file, no OCR) directly to
        factura_emitida/factura_recibida. Predates Phase 5's OCR upload
        endpoint (see POST /facturas/ocr, moved there specifically to avoid
        colliding with this pre-existing path -- see feature.md's documented
        deviation). Never recalculates cuota_iva/cuota_deducible -- those are
        computed client-side or defaulted, never derived here.
      operationId: crearFactura
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: "#/components/schemas/FacturaEmitidaInput"
                - $ref: "#/components/schemas/FacturaRecibidaInput"
      responses:
        "201":
          description: Invoice persisted
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /api/proceso/p04/facturas/ocr:
    post:
      summary: Upload invoice (PDF or image) for OCR extraction
      description: |
        Accepts a PDF or image file of an invoice. Runs Claude Vision OCR
        to extract invoice fields. Returns extracted data with confidence
        scores. Fields with confidence < 0.8 are flagged for user review
        (requires_review=true) -- the client-side FacturaReviewer component
        blocks progress until the user confirms or edits each such field.
        The file is stored in the `facturas` Supabase Storage bucket.
        Moved here from POST /facturas (Phase 1's existing JSON CRUD path)
        -- see feature.md's documented deviation.
      operationId: subirFacturaOcr
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, tipo]
              properties:
                file:
                  type: string
                  format: binary
                  description: PDF or image file of the invoice
                tipo:
                  type: string
                  enum: [emitida, recibida]
                  description: Whether this is an issued or received invoice
      responses:
        "200":
          description: OCR extraction result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OcrResult"
        "422":
          description: File type not supported or extraction failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /api/proceso/p04/mensaje/{proceso_id}:
    post:
      summary: Send a chat turn to the P04 conversational agent
      description: |
        Not in the original spec -- added because api-spec.yml never defined
        a chat-turn endpoint and ChatInterface needs one (see feature.md's
        documented deviation). Two payload shapes: `{"tipo": "texto",
        "contenido": str}` for a free-text message, or `{"tipo":
        "factura_confirmada", "factura": {...}}` to resolve a pending
        OCR-confirmation interrupt. A single free-text turn may run the
        graph all the way to the confirmar interrupt in one call -- the
        response's `pendiente` field tells the client what's next.
      operationId: enviarMensajeP04
      parameters:
        - name: proceso_id
          in: path
          required: true
          schema:
            type: string
          description: The LangGraph thread_id returned by POST /iniciar -- not a UUID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tipo]
              properties:
                tipo:
                  type: string
                  enum: [texto, factura_confirmada]
                contenido:
                  type: string
                  description: Required when tipo=texto
                factura:
                  type: object
                  description: Required when tipo=factura_confirmada
      responses:
        "200":
          description: Updated conversation state
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensajes:
                    type: array
                    items:
                      type: object
                      properties:
                        rol:
                          type: string
                          enum: [usuario, agente]
                        contenido:
                          type: string
                  pendiente:
                    type: string
                    enum: [none, revision_ocr, confirmacion]

  /api/proceso/p04/calcular:
    post:
      summary: Run fiscal engine and return M303 calculation
      description: |
        Takes all confirmed invoices for the period and runs the deterministic
        Python fiscal engine. Returns the full M303 calculation breakdown.
        Does NOT present to AEAT — only calculates.
        Saves the result to the presentacion table with estado=calculado.
      operationId: calcularP04
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [proceso_id, facturas_emitidas, facturas_recibidas]
              properties:
                proceso_id:
                  type: string
                  description: The thread_id returned by POST /iniciar -- not a UUID
                facturas_emitidas:
                  type: array
                  items:
                    $ref: "#/components/schemas/FacturaEmitidaInput"
                facturas_recibidas:
                  type: array
                  items:
                    $ref: "#/components/schemas/FacturaRecibidaInput"
      responses:
        "200":
          description: M303 calculation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResultadoM303"

  /api/proceso/p04/confirmar:
    post:
      summary: User confirms — enqueues RPA job for AEAT presentation
      description: |
        THE CRITICAL ENDPOINT. Called when the user explicitly confirms
        the presentation in the ConfirmacionModal.

        Records the confirmation with timestamp, user_id, and content hash.
        Changes estado to 'confirmado'.
        Enqueues the RPA job in ARQ/Redis.
        Returns immediately — the actual AEAT presentation is async.

        The client should then poll /estado or listen on Supabase Realtime
        for the presentado/error state change.
      operationId: confirmarP04
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [proceso_id]
              properties:
                proceso_id:
                  type: string
                  description: The thread_id returned by POST /iniciar -- not a UUID
                metodo_pago:
                  type: string
                  enum: [domiciliacion, nrc, tarjeta]
                  description: Payment method if result is a_ingresar
                iban:
                  type: string
                  nullable: true
                  description: Required if metodo_pago=domiciliacion
      responses:
        "202":
          description: Accepted — RPA job enqueued
          content:
            application/json:
              schema:
                type: object
                properties:
                  proceso_id:
                    type: string
                  rpa_job_id:
                    type: string
                  mensaje:
                    type: string
                    example: "Presentación en curso. Te notificaremos cuando esté lista."

  /api/proceso/p04/estado/{proceso_id}:
    get:
      summary: Poll process state
      description: |
        Returns the current state of the P04 process.
        Used for polling after confirmation, or to check if a previous
        session's process is still pending.
      operationId: estadoP04
      parameters:
        - name: proceso_id
          in: path
          required: true
          schema:
            type: string
          description: The thread_id returned by POST /iniciar -- not a UUID. Looked up against presentacion by (user_id, proceso, ejercicio, periodo), the same key notificar.py's upsert uses -- presentacion's own id column is a separate, internal UUID never exposed to the client.
      responses:
        "200":
          description: Current process state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcesoEstado"
        "404":
          description: Process not found for this user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /api/proceso/p04/justificante/{proceso_id}:
    get:
      summary: Get signed URL for PDF justificante download
      description: |
        Returns a signed Supabase Storage URL for the PDF justificante
        from the AEAT. Only available after estado=presentado.
        URL is valid for 60 minutes.
      operationId: justificanteP04
      parameters:
        - name: proceso_id
          in: path
          required: true
          schema:
            type: string
          description: The thread_id returned by POST /iniciar -- not a UUID (see /estado's parameter description)
      responses:
        "200":
          description: Signed download URL
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
                    description: Signed URL valid for 60 minutes
                  csv_aeat:
                    type: string
                    description: 16-char AEAT confirmation code
                  nrc:
                    type: string
                    nullable: true
        "404":
          description: Justificante not yet available or process not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /health:
    get:
      summary: Health check
      security: []
      operationId: health
      responses:
        "200":
          description: Service healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "ok"
                  db:
                    type: string
                    example: "ok"
                  redis:
                    type: string
                    example: "ok"